refactor(astra-runtime): modularize staging and checkpoint concerns
Split the 900-line monolithic lib.rs into focused, well-separated modules:

- src/utils.rs: Pure utilities (build_chunk_key, sanitize_name, parse_sequence, unix_time_ms)
- src/types.rs: Core data structures (StagingConfig, StageChunk, StageChunkPayload, MinioStagingConfig)
- src/staging/mod.rs: StageChunkStore trait definition
- src/staging/local.rs: LocalStageChunkStore implementation (filesystem backend)
- src/staging/minio.rs: MinioStageChunkStore implementation (MinIO/S3 backend) + bucket I/O helpers
- src/checkpoint.rs: Snapshot checkpoint persistence (LocalCheckpointStore, SnapshotCheckpointLedger)
- src/lib.rs: Crate root with re-exports, status(), and test suite

This refactor improves:
✓ Separation of concerns: each module has a single, clear responsibility
✓ Testability: logic is isolated and easier to unit test
✓ Maintainability: related code is co-located in dedicated modules
✓ Discoverability: developers can easily find where specific functionality lives
✓ Public API: unchanged, all downstream crates compile without modification

All 8 existing tests pass without modification.
